Reorganize modules and restore compat with core
