+++ Editor/RealityToolkit.Player.Editor.asmdef
@@ -2,6 +2,7 @@
     "name": "RealityToolkit.Player.Editor",
     "rootNamespace": "RealityToolkit.Player.Editor",
     "references": [
+        "GUID:f3241d040533491e8a1e2714b27c3111",
         "GUID:e67d30660ec243e4836aac191d3f36fb",
         "GUID:de07d9dfb71843b46aa2c474f73ca4cc",
         "GUID:b2d046948d6452a4b8485efc9ce0f88c",
