+++ Runtime/AssemblyInfo.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Reality Collective. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using System.Reflection;
+
+[assembly: AssemblyVersion("1.0.4")]
+[assembly: AssemblyTitle("com.realitytoolkit.player")]
+[assembly: AssemblyCompany("Reality Collective")]
+[assembly: AssemblyCopyright("Copyright (c) Reality Collective. All rights reserved.")]
