+++ Runtime/Bounds/PlayerOutOfBoundsFade.cs
@@ -1,11 +1,11 @@
-// Copyright (c) Reality Collective. All rights reserved.
+﻿// Copyright (c) Reality Collective. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using RealityCollective.ServiceFramework.Services;
-using RealityToolkit.Player.Interfaces;
+using RealityToolkit.Player.UX;
 using UnityEngine;
 
-namespace RealityToolkit.Player.UX
+namespace RealityToolkit.Player.Bounds
 {
     /// <summary>
     /// Uses the <see cref="IPlayerService.PlayerOutOfBounds"/>
