+++ Runtime/Bounds/PlayerOutOfBoundsTrigger.cs
@@ -1,9 +1,9 @@
-// Copyright (c) Reality Collective. All rights reserved.
+﻿// Copyright (c) Reality Collective. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using UnityEngine;
 
-namespace RealityToolkit.Player.UX
+namespace RealityToolkit.Player.Bounds
 {
     /// <summary>
     /// Put this component on a <see cref="GameObject"/> with a <see cref="Collider"/>
