+++ Runtime/Rigs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e398a07c523d9544bac264df6f057946
+guid: ed29b2040c8a9654cbae2a36ba2e7254
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
