+++ Runtime/Rigs/XRPlayerHead.cs
@@ -1,13 +1,12 @@
-// Copyright (c) Reality Collective. All rights reserved.
+﻿// Copyright (c) Reality Collective. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using RealityCollective.Extensions;
 using RealityCollective.ServiceFramework.Services;
-using RealityToolkit.Player.Interfaces;
-using RealityToolkit.Player.UX;
+using RealityCollective.Utilities.Extensions;
+using RealityToolkit.Player.Bounds;
 using UnityEngine;
 
-namespace RealityToolkit.Player
+namespace RealityToolkit.Player.Rigs
 {
     [RequireComponent(typeof(Rigidbody))]
     [RequireComponent(typeof(SphereCollider))]
