+++ Runtime/Rigs/XRPlayerRig.cs
@@ -1,16 +1,14 @@
 ﻿// Copyright (c) Reality Collective. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using RealityCollective.Extensions;
 using RealityCollective.ServiceFramework.Services;
-using RealityToolkit.Player.Definitions;
-using RealityToolkit.Player.Interfaces;
+using RealityCollective.Utilities.Extensions;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SpatialTracking;
 using UnityEngine.XR;
 
-namespace RealityToolkit.Player
+namespace RealityToolkit.Player.Rigs
 {
     /// <summary>
     /// The default <see cref="IXRPlayerRig"/> implementation.
