+++ Runtime/UX/CameraFade.cs
@@ -1,7 +1,7 @@
-// Copyright (c) Reality Collective. All rights reserved.
+﻿// Copyright (c) Reality Collective. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using RealityCollective.Extensions;
+using RealityCollective.Utilities.Extensions;
 using System.Collections;
 using UnityEngine;
